Added dialoque feature for creating new notes
Used the dialoguer crate to interact and ask the user for the name
which he didn't provide and if he wants to edit it in either
notepad or gedit.

```bash
alina note new

alina note list
```

In the future macOS should also be supported but I don't know the
std editor.
Additonaly the user should be able to use his editor of choice and
configure it in the config file or through the cli with a config
command.