Added iterator implementation for config and new options to config
Added an iterator implementation to the config struct and therefor
an `ConfigIterator` struct. Reason for that is mainly because I
wanted to print out alle keys and values so the user is able to
see which options he set and maybe gets an idea of what he or she
is able to configure.

Additionally I added an `Editor` option to config which is used
for opening files and editing them.